% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper.R
\name{add_grid_lines_piper}
\alias{add_grid_lines_piper}
\title{add_grid_lines_piper}
\usage{
add_grid_lines_piper(color = NULL, Size = NULL)
}
\arguments{
\item{color}{A character string specifying the color of the grid lines in 
Piper plot}

\item{Size}{A numeric value specifying the size of the grid labels in a 
Piper plot}
}
\value{
This function returns a list with the geom components used to create 
a grid in a Piper diagram
}
\description{
Function to add the grid lines to a Piper plot
}
\seealso{
Other piper functions: \code{\link{add_labels_piper}},
  \code{\link{anion_cation_transform}},
  \code{\link{ggplot_piper}},
  \code{\link{piper_classification}},
  \code{\link{piper_transform}}, \code{\link{plot_piper}},
  \code{\link{transform_piper_data}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{piper functions}
