% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multirectangular.R
\name{add_labels_multirectangular}
\alias{add_labels_multirectangular}
\title{add_labels_multirectangular}
\usage{
add_labels_multirectangular(color = NULL, Size = NULL)
}
\arguments{
\item{color}{A character string specifying the color of the labels}

\item{Size}{A numeric value specifying the size of the text label}
}
\value{
This function returns a geom_text to be added to the multirectangular
 diagram
}
\description{
Function to add the labels to a multirectangular diagram
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{multirectanguar functions}
