% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper.R
\name{add_labels_piper}
\alias{add_labels_piper}
\title{add_labels_piper}
\usage{
add_labels_piper(labels, middle = TRUE, color = NULL, Size = NULL)
}
\arguments{
\item{labels}{A character vector with the labels to be added to the 
Piper diagram}

\item{middle}{A logical flag indicating if the labels must be located on the 
middle part of the axis.}

\item{color}{A character string specifying the color of the labels}

\item{Size}{A numeric value specifying the size of the text label}
}
\value{
This function returns a geom_text to be added to the Piper diagram
}
\description{
Function to add the labels to a ternary diagram
}
\seealso{
Other piper functions: \code{\link{add_grid_lines_piper}},
  \code{\link{anion_cation_transform}},
  \code{\link{ggplot_piper}},
  \code{\link{piper_classification}},
  \code{\link{piper_transform}}, \code{\link{plot_piper}},
  \code{\link{transform_piper_data}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{piper functions}
