% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary.R
\name{add_labels_ternary}
\alias{add_labels_ternary}
\title{add_labels_ternary}
\usage{
add_labels_ternary(labels, middle = TRUE, color = NULL, Size = NULL)
}
\arguments{
\item{labels}{A character vector with the labels to be added to the 
ternary plot}

\item{middle}{A logical flag indicating if the labels must be located on the 
middle part of the axis.}

\item{color}{A character string specifying the color of the labels}

\item{Size}{A numeric value specifying the size of the text label}
}
\value{
This function returns a geom_text to be added to the
}
\description{
Function to add the labels to a ternary diagram
}
\seealso{
Other ternary functions: \code{\link{add_grid_lines_ternary}},
  \code{\link{ggplot_ternary}}, \code{\link{plot_ternary}},
  \code{\link{ternary_transform}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{ternary functions}
