% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper.R
\name{anion_cation_transform}
\alias{anion_cation_transform}
\title{anion_cation_transform}
\usage{
anion_cation_transform(Ca, Mg, Na, K, HCO3, CO3, Cl, SO4, name = NULL)
}
\arguments{
\item{Ca}{A numeric vector with the Ca concentration}

\item{Mg}{A numeric vector with the Mg concentration}

\item{Na}{A numeric vector with the Na concentration}

\item{K}{A numeric vector with the K concentration}

\item{HCO3}{A numeric vector with the HCO3 concentration}

\item{CO3}{A numeric vector with the CO3 concentration}

\item{Cl}{A numeric vector with the Cl concentration}

\item{SO4}{A numeric vector with the SO4 concentration}

\item{name}{A numeric vector with the Ca concentration}
}
\value{
A list with the following entries:
\itemize{
\item cat_x,cat_y: numeric vectors with the x and y coordinates used to create the
ternary diagram of the cations.
\item an_x, an_ynumeric vectors with the x and y coordinates used to create the ternary
 diagram of the anions.
\item d_x, d_y
\item  meqL: A numeric vector with the meqL of the input concentrations.
\item sumanions: Sum of the meqL of the anions
\item anions: A matrix with the values of the normalized meqL of the anions.
\item cations: A matrix with the values of the normalized meqL of the anions.
}
}
\description{
Function to calculate the coordinates of anions and cations used in the Piper diagram.
}
\seealso{
Other piper functions: \code{\link{add_grid_lines_piper}},
  \code{\link{add_labels_piper}},
  \code{\link{ggplot_piper}},
  \code{\link{piper_classification}},
  \code{\link{piper_transform}}, \code{\link{plot_piper}},
  \code{\link{transform_piper_data}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{piper functions}
