% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_utilities.R
\name{anomaly_concentration_area_method}
\alias{anomaly_concentration_area_method}
\title{anomaly_concentration_area_method}
\usage{
anomaly_concentration_area_method(gdata, int.method = c("akima", "idw",
  "kriging"))
}
\arguments{
\item{gdata}{A geochemical dataset object}

\item{int.method}{A character string with the interpolation method. Currently three methods
are supported:
\itemize{
\item akima
\item idw
\item kriging
}}
}
\value{
This function returns a list with the following entries
}
\description{
Identification of anomalous areas using the concentration-area approach
}
\seealso{
Other anomaly functions: \code{\link{anomaly_graphical_method}},
  \code{\link{anomaly_statistical_methods}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
\concept{anomaly functions}
