% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_utilities.R
\name{anomaly_graphical_method}
\alias{anomaly_graphical_method}
\title{anomaly_graphical_method}
\usage{
anomaly_graphical_method(gdata)
}
\arguments{
\item{gdata}{A geochemical dataset object}
}
\value{
This function returns a ggplot2 object with the QQplots of each major ion
}
\description{
Function to create the QQplots required in the identification of homogeneous statistical
groups.
}
\seealso{
Other anomaly functions: \code{\link{anomaly_concentration_area_method}},
  \code{\link{anomaly_statistical_methods}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
\concept{anomaly functions}
