% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_utilities.R
\name{anomaly_statistical_methods}
\alias{anomaly_statistical_methods}
\title{anomaly_statistical_methods}
\usage{
anomaly_statistical_methods(gdata, method = c("mean", "quantile"))
}
\arguments{
\item{gdata}{A geochemical dataset object}

\item{method}{A character string with the method to define the threshold}
}
\value{
This function returns a list with the following entries
}
\description{
Function to calculate the anomaly threshold using two statistical methods:
1. mean + 2*std
2. Quantile 90
}
\seealso{
Other anomaly functions: \code{\link{anomaly_concentration_area_method}},
  \code{\link{anomaly_graphical_method}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
