% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geochemical_dataset_utilities.R
\name{calculate_convervative_mixing}
\alias{calculate_convervative_mixing}
\title{calculate_conservative_mixing}
\usage{
calculate_convervative_mixing(gdata, fresh.water = NULL,
  sea.water = NULL)
}
\arguments{
\item{gdata}{A data.frame with the concentrations of the major ions}

\item{fresh.water}{The chemical composition of fresh water}

\item{sea.water}{The chemical composition of sea water}
}
\value{
This function returns a list with the following entries:
\itemize{
\item f.sea: A vector with the mixing ratio
\item mix: Matrix with the results of conservative mixing between fresh and sea water (in mmol/L)
\item react: Matrix with the concentrations due to reactions (in mmol/L)
}
}
\description{
Function to calculate the mixing ratio of a perfect-conservative mixture between fresh and
sea water from a given geochemical dataset.
}
\seealso{
Other base functions: \code{\link{convert_meql}},
  \code{\link{geochemical_dataset}},
  \code{\link{plot.geochemical_dataset}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
\concept{base functions}
