% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geothermometers_utilities.R
\name{calculate_geothermometers}
\alias{calculate_geothermometers}
\title{calculate_geothermometers}
\usage{
calculate_geothermometers(gdata)
}
\arguments{
\item{gdata}{A geochemical_dataset object}
}
\value{
This function returns a data.frame with the temperatures
}
\description{
Function to calculate several geothermometers
}
\seealso{
Other geothermometer functions: \code{\link{Fournier.Potter.geothermometer}},
  \code{\link{K.Mg.geothermometer}},
  \code{\link{Li.Mg.geothermometer}},
  \code{\link{Na.K.Ca.geothermometer}},
  \code{\link{Na.K.geothermometers}},
  \code{\link{silica.geothermometers}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
\concept{geothermometer functions}
