% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper_modified.R
\name{classification_reactions_piper_mode}
\alias{classification_reactions_piper_mode}
\title{classification_reactions_piper_mod}
\usage{
classification_reactions_piper_mode(dx, dy)
}
\arguments{
\item{dx}{A numeric vector with the increments in x}

\item{dy}{A numeric vector with the increments in y}
}
\value{
This function returns a matrix with the given contribution of each type of chemical reaction
}
\description{
Function to calculate the contribution of each chemical reaction to the chemical composition
of a given sample
}
\seealso{
Other modified_piper functions: \code{\link{modified_piper_transform}},
  \code{\link{plot_modified_piper}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
\concept{modified_piper functions}
