% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durov.R
\name{durov_transform}
\alias{durov_transform}
\title{durov_transform}
\usage{
durov_transform(Ca, Mg, Na, K, HCO3, CO3, Cl, SO4, pH, TDS)
}
\arguments{
\item{Ca}{A numeric vector with the Ca concentration}

\item{Mg}{A numeric vector with the Mg concentration}

\item{Na}{A numeric vector with the Na concentration}

\item{K}{A numeric vector with the K concentration}

\item{HCO3}{A numeric vector with the HCO3 concentration}

\item{CO3}{A numeric vector with the CO3 concentration}

\item{Cl}{A numeric vector with the Cl concentration}

\item{SO4}{A numeric vector with the SO4 concentration}

\item{pH}{A numeric vector with the pH concentration}

\item{TDS}{A numeric vector with the TDS concentration}
}
\value{
A list with the following entries
\itemize{
\item cx_mnsquare, cy_mnsquare
}
}
\description{
Calculate the coordinates in the Durov plot.
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
