% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durov.R
\name{durov_transform1}
\alias{durov_transform1}
\title{durov_transform1}
\usage{
durov_transform1(gdata)
}
\arguments{
\item{gdata}{A geochemical_dataset object}
}
\value{
A data.frame with the coordinates
}
\description{
Function to calculate the coordinates of a given composition in a Durov Diagram.
}
\seealso{
Other durov functions: \code{\link{ggplot_durov}},
  \code{\link{plot_durov}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{durov functions}
