% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geothermometers_utilities.R
\name{Fournier.Potter.geothermometer}
\alias{Fournier.Potter.geothermometer}
\title{Fournier.Potter.geothermometer}
\usage{
Fournier.Potter.geothermometer(SiO2, Temp)
}
\arguments{
\item{SiO2}{A numeric vector with the SiO2 concentration}

\item{Temp}{A numeric vector with the temperature of the samples}
}
\value{
This function returns a data.frame with the temperatures
}
\description{
This function calculates the Fournier-Potter SiO2 geothermometer
}
\seealso{
Other geothermometer functions: \code{\link{K.Mg.geothermometer}},
  \code{\link{Li.Mg.geothermometer}},
  \code{\link{Na.K.Ca.geothermometer}},
  \code{\link{Na.K.geothermometers}},
  \code{\link{calculate_geothermometers}},
  \code{\link{silica.geothermometers}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
\concept{geothermometer functions}
