% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geochemical_dataset_utilities.R
\name{geochemical_dataset}
\alias{geochemical_dataset}
\title{geochemical_dataset}
\usage{
geochemical_dataset(name = "GeochemicalDataset", data)
}
\arguments{
\item{name}{Name of the geochemical dataset}

\item{data}{Data.frame with the concentration of anions and cations (see details)}
}
\value{
An object of the geochemical_dataset class with the following member variables:
\itemize{
\item name: Name of the geochemical dataset
\item dataset: data.frame with the corresponding measured concentrations
\item meqL: data.frame with the meqL of the measured concentrations
\item anions: data.frame with the meqL of the major anions
\item cations: data.frame with the meqL of the major cations
\item meqL.min: Numeric vector with the minimum values of the meqL
\item meqL.max: Numeric vector with the maximum values of the meqL
\item meqL.schoeller: data.frame with the concentrations of the major ions organized to
create tehe Schoeller plot.
}
}
\description{
Function to create a geochemical_dataset.
}
\details{
A geochemical dataset is a data.frame with the values of the concentrations of the anions
and cations obtained during hydrogeochemical sampling. Internally this object calculates the
meqL of the measured concentrations of the major ions. The concentrations of the ions must be
specified in mg/L.
}
\seealso{
Other base functions: \code{\link{convert_meql}},
  \code{\link{plot.geochemical_dataset}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
