% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durov.R
\name{ggplot_durov}
\alias{ggplot_durov}
\title{ggplot_durov}
\usage{
ggplot_durov()
}
\value{
A ggplot with the Durov diagram
}
\description{
Function
}
\seealso{
Other durov functions: \code{\link{durov_transform1}},
  \code{\link{plot_durov}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
