% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giggenbach_utilities.R
\name{ggplot_giggenbach}
\alias{ggplot_giggenbach}
\title{ggplot_giggenbach}
\usage{
ggplot_giggenbach()
}
\value{
This function returns the base of the Giggenbach's geothermometer plot
}
\description{
Function to create the base of the Giggenbach's geothermometer plot
}
\seealso{
Other giggenbach functions: \code{\link{add_lines_giggenbach}},
  \code{\link{plot_giggenbach}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{giggenbach functions}
