% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maucha.R
\name{ggplot_maucha}
\alias{ggplot_maucha}
\title{ggplot_maucha}
\usage{
ggplot_maucha()
}
\value{
This function returns a gggplot2 object with the template of the Maucha plot.
}
\description{
Function to create the template of the Maucha plot.
}
\seealso{
Other maucha functions: \code{\link{maucha_transform}},
  \code{\link{plot_maucha}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
