% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper_modified.R
\name{ggplot_modified_piper}
\alias{ggplot_modified_piper}
\title{ggplot_modified_piper}
\usage{
ggplot_modified_piper()
}
\value{
This function returns the base of the modified Piper plot
}
\description{
Function to create the base of the modified Piper plot as proposed by Appelo and Postma.
}
\seealso{
Other modified_piper functions: \code{\link{add_lines_modified_piper}},
  \code{\link{classification_reactions_piper_mode}},
  \code{\link{modified_piper_transform}},
  \code{\link{plot_modified_piper}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{modified_piper functions}
