% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multirectangular.R
\name{ggplot_multirectangular}
\alias{ggplot_multirectangular}
\title{ggplot_multirectangular}
\usage{
ggplot_multirectangular()
}
\value{
A ggplot2 plot with the template of a multirectangular diagram.
}
\description{
Function to create the template of a multirectangular diagram.
}
\seealso{
Other multirectangular functions: \code{\link{multirectangular_transform}},
  \code{\link{plot_multirectangular}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
