% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper.R
\name{ggplot_piper}
\alias{ggplot_piper}
\title{ggplot_piper}
\usage{
ggplot_piper()
}
\value{
A ggplot2 object
}
\description{
Function to create the ternary diagrams that compose the Piper diagram
}
\seealso{
Other piper functions: \code{\link{add_grid_lines_piper}},
  \code{\link{add_labels_piper}},
  \code{\link{anion_cation_transform}},
  \code{\link{piper_classification}},
  \code{\link{piper_transform}}, \code{\link{plot_piper}},
  \code{\link{transform_piper_data}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{piper functions}
