% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial.R
\name{ggplot_radial}
\alias{ggplot_radial}
\title{ggplot_radial}
\usage{
ggplot_radial()
}
\value{
A ggplot2 with the template of a Radial diagram
}
\description{
Function to create the template of a radial diagram
}
\seealso{
Other radial functions: \code{\link{plot_radial}},
  \code{\link{radial_transform}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
