% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiff.R
\name{ggplot_stiff}
\alias{ggplot_stiff}
\title{ggplot_stiff}
\usage{
ggplot_stiff()
}
\value{
A ggplot2 plot with the Stiff diagram template
}
\description{
Function to create a Stiff diagram
}
\seealso{
Other stiff functions: \code{\link{plot_stiff}},
  \code{\link{stiff_transform}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
