% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giggenbach_utilities.R
\name{giggenbach_ternary_transform}
\alias{giggenbach_ternary_transform}
\title{giggenbach_ternary_transform}
\usage{
giggenbach_ternary_transform(gdata)
}
\arguments{
\item{gdata}{A geochemical_dataset object}
}
\value{
This function returns a data.frame with the following entries
\itemize{
\item cNa: concentrations of Na divided by 1000
\item cK: concentration of K divided by 100
\item cMg: square root of concentration of Mg
}
}
\description{
This function calculates the coordinates of the ternary plot proposed by Giggenbach (1988)
to identify the samples that can be used in the geothermometer calculations
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
\concept{giggenbach_functions}
