% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_utilities.R,
%   R/compostional_ilr_utilities.R, R/durov.R,
%   R/geochemical_dataset_utilities.R, R/geothermometers_utilities.R,
%   R/giggenbach_utilities.R, R/maucha.R, R/mixing_utilities.R,
%   R/multirectangular.R, R/piper.R, R/piper_modified.R, R/radial.R, R/stiff.R,
%   R/ternary.R
\docType{package}
\name{GQAnalyzer}
\alias{GQAnalyzer}
\alias{GQAnalyzer-package}
\title{GQAnalyzer: A package for Hydrogeochemical plotting}
\description{
This package provides functions to create the following plots:
\itemize{
\item Piper plot
\item Stiff diagrams
\item Durov plot
\item Ternary
\item Radial
\item Multirectangular
\item Maucha
}
}
\section{Anomaly Utilities}{


This section includes functions used in the identification of anomalies in groundwater quality data.

This section includes
}

\section{Compositional-ILR Diagram}{


The functions defined to create a diagram with an ILR transform are:
ilr_transform, plot_ilr
}

\section{Durov Diagram}{


The Durov diagram is a composite diagram that includes a central rectangular diagram of

This section includes the definition of three functions used to create the Durov diagram:
durov_transform, durov_transform1, ggplot_durov
}

\section{Geothermometers}{


The functions defined to calculate the temperatures
}

\section{Giggenbach Diagrams}{


The functions defined to create the Giggenbach diagrams are:
}

\section{Maucha Diagram}{


The functions defined to create a Multirectangular diagram are:
maucha_transform, ggplot_maucha
}

\section{Mixing Utilities}{


This section is devoted to functions used in the study of mixing processes from hydrogeochenmical information.

This section includes the definition of three functions used to create the Durov diagram:
m3_model
}

\section{Multirectangular Diagram}{


The functions defined to create a Multirectangular diagram are:
multirectangular_transform, ggplot_multirectangular
}

\section{Piper Diagram}{


The functions defined to create a Piper diagram are:
anion_cation_tranform, transform_piper_data, ggplot_piper
}

\section{Modified Piper Diagram}{


The functions defined to create a Modified Piper diagram are:
transform_modified_piper_data, ggplot_modified_piper
}

\section{Radial Diagram}{


The functions defined to create a Radial diagram are:
radial_tranform, ggplot_radial
}

\section{Stiff Diagram}{


The functions defined to create a Stiff diagram are:
stiff_tranform, ggplot_stiff
}

\section{Ternary Plot}{


The functions defined to create a Ternary plot are:
ternary_transform, ggplot_ternary
}

