% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compostional_ilr_utilities.R
\name{ilr_transform}
\alias{ilr_transform}
\title{ilr_transform}
\source{
Shelton, Jenna L., Engle, Mark A., Buccianti, Antonella, Blondes, Madalyn S. (2018).
The isometric log-ratio (ilr)-ion plot: A proposed alternative to the Piper diagram.
ournal of Geochemical Exploration, Vol 190, Issue March.
}
\usage{
ilr_transform(gdata)
}
\arguments{
\item{gdata}{A geochemical_dataset object}
}
\value{
A data.frame with the coordinates
}
\description{
Function to calculate the coordinates of a geochemical_dataset in a multirectangular diagram.
}
\seealso{
Other ilr functions: \code{\link{plot_ilr}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{ilr functions}
