% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geothermometers_utilities.R
\name{Li.Mg.geothermometer}
\alias{Li.Mg.geothermometer}
\title{Mg.Li.geothermometer}
\usage{
Li.Mg.geothermometer(Li, Mg, Temp)
}
\arguments{
\item{Li}{A numerical vector with K concentrations}

\item{Mg}{A numerical vector with Mg concentrations}

\item{Temp}{A numerical vector with temperature concentrations}
}
\value{
This function returns a data.frame with the temperatures
}
\description{
This function calculates tshe Na.K.Ca geothermometer
}
\seealso{
Other geothermometer functions: \code{\link{Fournier.Potter.geothermometer}},
  \code{\link{K.Mg.geothermometer}},
  \code{\link{Na.K.Ca.geothermometer}},
  \code{\link{Na.K.geothermometers}},
  \code{\link{calculate_geothermometers}},
  \code{\link{silica.geothermometers}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
\concept{geothermometer functions}
