% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maucha.R
\name{maucha_transform}
\alias{maucha_transform}
\title{maucha_transform}
\usage{
maucha_transform(gdata)
}
\arguments{
\item{gdata}{A geochemical_dataset object}
}
\value{
This function returns a data.frame with the following variables:
\itemize{
\item xcirc
\item ycirc
}
}
\description{
Function to calculate the coordinates of a given sample in a Maucha plot
}
\seealso{
Other maucha functions: \code{\link{ggplot_maucha}},
  \code{\link{plot_maucha}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
