% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_utilities.R
\name{mix_model}
\alias{mix_model}
\title{mix_model}
\usage{
mix_model(gdata.gd, end.members.gd)
}
\arguments{
\item{gdata.gd}{A geochemical_dataset}

\item{end.members.gd}{A geochemical_dataset}
}
\value{
A numeric vector
}
\description{
Function to estimate mixing ratios with uncertain end.members
}
\seealso{
Other mixing functions: \code{\link{constrained_lm}},
  \code{\link{m3_mixing_model}},
  \code{\link{plot_m3_mixing_results}},
  \code{\link{select_end_members}},
  \code{\link{uncertainty_m3_mixing_model}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{mixing functions}
