% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper_modified.R
\name{modified_piper_transform}
\alias{modified_piper_transform}
\title{modified_piper_transform}
\usage{
modified_piper_transform(gdata)
}
\arguments{
\item{gdata}{A geochemical_dataset}
}
\value{
This function returns a data.frame with the following variables:
\itemize{
\item cat_x,cat_y: coordinates of the cations in the ternary diagram (scaled between 0 and 1).
\item an_x,an_y: coordiantes of the anions in the ternary diagram (scaled between 0 and 1).
\item d_x,d_y: coordinates of the samples in the central diamond (scaled between 0 and 1).
\item k: Matrix with the increments in the coordinates of the sample tail in the central diamond.
}
}
\description{
Function to calculate the coordinates of set of samples in a piper diagram.
}
\seealso{
Other modified_piper functions: \code{\link{add_lines_modified_piper}},
  \code{\link{classification_reactions_piper_mode}},
  \code{\link{ggplot_modified_piper}},
  \code{\link{plot_modified_piper}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{modified_piper functions}
