% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multirectangular.R
\name{multirectangular_transform}
\alias{multirectangular_transform}
\title{multirectangular_transform}
\usage{
multirectangular_transform(gdata)
}
\arguments{
\item{gdata}{A geochemical_dataset object}
}
\value{
A data.frame with the coordinates
}
\description{
Function to calculate the coordinates of a geochemical_dataset in a multirectangular diagram.
}
\seealso{
Other multirectangular functions: \code{\link{ggplot_multirectangular}},
  \code{\link{plot_multirectangular}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
