% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geothermometers_utilities.R
\name{Na.K.geothermometers}
\alias{Na.K.geothermometers}
\title{Na.K.geothermometers}
\usage{
Na.K.geothermometers(Na, K, Temp)
}
\arguments{
\item{Na}{A numerical vector with Na concentrations}

\item{K}{A numerical vector with K concentrations}

\item{Temp}{A numerical vector with temperature concentrations}
}
\value{
This function returns a data.frame with the temperatures
}
\description{
This function calculates several Na-K geothermometers
}
\seealso{
Other geothermometer functions: \code{\link{Fournier.Potter.geothermometer}},
  \code{\link{K.Mg.geothermometer}},
  \code{\link{Li.Mg.geothermometer}},
  \code{\link{Na.K.Ca.geothermometer}},
  \code{\link{calculate_geothermometers}},
  \code{\link{silica.geothermometers}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
\concept{geothermometer functions}
