% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper.R
\name{piper_classification}
\alias{piper_classification}
\title{piper_classification}
\usage{
piper_classification(d_x, d_y)
}
\arguments{
\item{d_x}{A numeric vector with the x coordinates of the samples inside the diamond of the Piper diagram.}

\item{d_y}{A numeric vector with the x coordinates of the samples inside the diamond of the Piper diagram.}
}
\value{
This function returns a character vector with the corresponding classification
}
\description{
Function to classify the groundwater samples according to the Piper diagram
}
\seealso{
Other piper functions: \code{\link{add_grid_lines_piper}},
  \code{\link{add_labels_piper}},
  \code{\link{anion_cation_transform}},
  \code{\link{ggplot_piper}},
  \code{\link{piper_transform}}, \code{\link{plot_piper}},
  \code{\link{transform_piper_data}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
\concept{piper functions}
