% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper.R
\name{piper_transform}
\alias{piper_transform}
\title{piper_transform}
\usage{
piper_transform(gdata)
}
\arguments{
\item{gdata}{A geochemical_dataset}
}
\value{
This function returns a data.frame with the following variables:
\itemize{
\item cat_x,cat_y: coordinates of the cations in the ternary diagram (scaled between 0 and 1).
\item an_x,an_y: coordiantes of the anions in the ternary diagram (scaled between 0 and 1).
\item d_x,d_y: coordinates of the samples in the central diamond (scaled between 0 and 1).
}
}
\description{
Function to calculate the coordinates of set of samples in a piper diagram.
}
\seealso{
Other piper functions: \code{\link{add_grid_lines_piper}},
  \code{\link{add_labels_piper}},
  \code{\link{anion_cation_transform}},
  \code{\link{ggplot_piper}},
  \code{\link{piper_classification}},
  \code{\link{plot_piper}},
  \code{\link{transform_piper_data}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{piper functions}
