% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geochemical_dataset_utilities.R
\name{plot.geochemical_dataset}
\alias{plot.geochemical_dataset}
\title{plot.geochemical_dataset}
\usage{
\method{plot}{geochemical_dataset}(x, y = NULL, ..., type = c("piper",
  "ternary", "durov", "schoeller", "stiff", "multirectangular", "radial"),
  measure = c("conc", "meql"), vars = NULL, color = NULL, Size = NULL)
}
\arguments{
\item{x, y}{A geochemical_dataset object}

\item{...}{Additional parameters to the plot function}

\item{type}{A character string with the type of hydrogeochemical plot to be created.
Currently the following plots are supported:
\itemize{
\item piper
\item ternary
\item durov
\item schoeller
\item stiff
\item multirectangular
\item radial
\item maucha
}}

\item{measure}{A character variable specifying the type of measure to be used in the plot.
Currently the types supported include:
\itemize{
\item 'conc': concentrations
\item 'meql': miliequivalent
}}

\item{vars}{Character vector. Variables to use in the plot. In the case of the Durov plot, these variables
are already defined. Used only for compatibility with the plot function.}

\item{color}{Character variable that specifies the variable to color the data inside the plot.}

\item{Size}{Character variable that specifies the variable to define the size of the data inside the plot.}
}
\value{
This function returns a ggplot2 object with the corresponding hydrogeochemical plot.
}
\description{
Generic function to create different hydrogeochemical plots
}
\seealso{
Other base functions: \code{\link{calculate_convervative_mixing}},
  \code{\link{convert_meql}},
  \code{\link{geochemical_dataset}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
