% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durov.R
\name{plot_durov}
\alias{plot_durov}
\title{plot_durov}
\usage{
plot_durov(x, measure = c("conc", "meql"), vars = NULL, color = NULL,
  Size = NULL, labels = NULL, additional.args = NULL)
}
\arguments{
\item{x}{A geochemical_dataset object}

\item{measure}{A character variable specifying the type of measure to be used in the plot.
Currently the types supported include:
\itemize{
\item 'conc': concentrations
\item 'meql': miliequivalent
}}

\item{vars}{Character vector. Variables to use in the plot. In the case of the Durov plot, these variables
are already defined. Used only for compatibility with the plot function.}

\item{color}{Character variable that specifies the variable to color the data inside the plot.}

\item{Size}{Character variable that specifies the variable to define the size of the data inside the plot.}

\item{labels}{Character variable that specifies the labels to be used in the current plot}

\item{additional.args}{A list with additional arguments}
}
\value{
This function returns a ggplot2 object with the Durov plot.
}
\description{
Function to create the Durov plot
}
\seealso{
Other durov functions: \code{\link{durov_transform1}},
  \code{\link{ggplot_durov}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
\concept{durov functions}
