% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giggenbach_utilities.R
\name{plot_giggenbach}
\alias{plot_giggenbach}
\title{plot_giggenbach}
\usage{
plot_giggenbach(x, measure = c("conc", "meql"), vars = NULL,
  color = NULL, Size = NULL)
}
\arguments{
\item{x}{A geochemical_dataset object}

\item{measure}{A character variable specifying the type of measure to be used in the plot.
Currently the types supported include:
\itemize{
\item 'conc': concentrations
\item 'meql': miliequivalent
}}

\item{vars}{Character vector. Variables to use in the plot. In the case of the Durov plot, these variables
are already defined. Used only for compatibility with the plot function.}

\item{color}{Character variable that specifies the variable to color the data inside the plot.}

\item{Size}{Character variable that specifies the variable to define the size of the data inside the plot.}
}
\value{
This function returns a ggplot2 object with the Giggenbach ternary plot.
}
\description{
Function to create the ternary plot proposed by Giggenbach(1988) to identify the samples
to be used in geothermometer calculations
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{giggenbach functions}
