% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compostional_ilr_utilities.R
\name{plot_ilr}
\alias{plot_ilr}
\title{plot_ilr}
\usage{
plot_ilr(x, measure = c("conc", "meql"), vars = NULL, color = NULL,
  Size = NULL)
}
\arguments{
\item{x}{A geochemical_dataset object}

\item{measure}{A character variable specifying the type of measure to be used in the plot.
Currently the types supported include:
\itemize{
\item 'conc': concentrations
\item 'meql': miliequivalent
}}

\item{vars}{Character vector. Variables to use in the plot. In the case of the Durov plot, these variables
are already defined. Used only for compatibility with the plot function.}

\item{color}{Character variable that specifies the variable to color the data inside the plot.}

\item{Size}{Character variable that specifies the variable to define the size of the data inside the plot.}
}
\value{
This function returns a ggplot2 object with the Durov plot.
}
\description{
This function creates the compositional plot
}
\seealso{
Other ilr functions: \code{\link{ilr_transform}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
\concept{ilr functions}
