% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multirectangular.R
\name{plot_multirectangular}
\alias{plot_multirectangular}
\title{plot_multirectangular}
\usage{
plot_multirectangular(x, measure = c("conc", "meql"), vars = NULL,
  color = NULL, Size = NULL)
}
\arguments{
\item{x}{A geochemical_dataset object}

\item{measure}{A character variable specifying the type of measure to be used in the plot.
Currently the types supported include:
\itemize{
\item 'conc': concentrations
\item 'meql': miliequivalent
}}

\item{vars}{Character vector. Variables to use in the plot. In the case of the Durov plot, these variables
are already defined. Used only for compatibility with the plot function.}

\item{color}{Character variable that specifies the variable to color the data inside the plot.}

\item{Size}{Character variable that specifies the variable to define the size of the data inside the plot.}
}
\value{
This function returns a ggplot2 object with the Durov plot.
}
\description{
Function to create the Multirectangular plot
}
\references{
Ahmad, N., Sen, Z., & Ahmad, M. (2003). Ground Water Quality Assessment Using Multi-Rectangular Diagrams.
Ground Water, 41(6), 828–832. http://doi.org/10.1111/j.1745-6584.2003.tb02423.x
}
\seealso{
Other multirectangular functions: \code{\link{ggplot_multirectangular}},
  \code{\link{multirectangular_transform}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
