% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper.R
\name{plot_piper}
\alias{plot_piper}
\title{plot_piper}
\usage{
plot_piper(x, measure = c("conc", "meql"), vars = NULL, color = NULL,
  Size = NULL)
}
\arguments{
\item{x}{A geochemical_dataset object}

\item{measure}{A character variable specifying the type of measure to be used in the plot.
Currently the types supported include:
\itemize{
\item 'conc': concentrations
\item 'meql': miliequivalent
}}

\item{vars}{Character vector. Variables to use in the plot. In the case of the Durov plot, these variables
are already defined. Used only for compatibility with the plot function.}

\item{color}{Character variable that specifies the variable to color the data inside the plot.}

\item{Size}{Character variable that specifies the variable to define the size of the data inside the plot.}
}
\value{
This function returns a ggplot2 object with the Durov plot.
}
\description{
Function to create the Piper plot
}
\seealso{
Other piper functions: \code{\link{anion_cation_transform}},
  \code{\link{ggplot_piper}},
  \code{\link{piper_classification}},
  \code{\link{piper_transform}},
  \code{\link{transform_piper_data}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
\concept{piper functions}
