% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiff.R
\name{plot_stiff}
\alias{plot_stiff}
\title{plot_stiff}
\usage{
plot_stiff(x, measure = c("conc", "meql"), vars = NULL, color = NULL,
  Size = NULL)
}
\arguments{
\item{x}{A geochemical_dataset object}

\item{measure}{A character variable specifying the type of measure to be used in the plot.
Currently the types supported include:
\itemize{
\item 'conc': concentrations
\item 'meql': miliequivalent
}}

\item{vars}{Character vector. Variables to use in the plot. In the case of the Durov plot, these variables
are already defined. Used only for compatibility with the plot function.}

\item{color}{Character variable that specifies the variable to color the data inside the plot.}

\item{Size}{Character variable that specifies the variable to define the size of the data inside the plot.}
}
\value{
This function returns a ggplot2 object with the Durov plot.
}
\description{
Function to create a Stiff plot for a given sample
}
\seealso{
Other stiff functions: \code{\link{ggplot_stiff}},
  \code{\link{stiff_transform}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
