% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial.R
\name{radial_transform}
\alias{radial_transform}
\title{radial_transform}
\usage{
radial_transform(gdata)
}
\arguments{
\item{gdata}{A geochemical_dataset object}
}
\value{
A data.frame with the corresponding coordinates
}
\description{
Function to calculate the coordinates of a geochemical_dataset in a Radial Diagram
}
\seealso{
Other radial functions: \code{\link{ggplot_radial}},
  \code{\link{plot_radial}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
