% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_utilities.R
\name{select_end_members}
\alias{select_end_members}
\title{select_end_members}
\usage{
select_end_members(gdata)
}
\arguments{
\item{gdata}{A geochemical_dataset object}
}
\value{
This function returns
}
\description{
Function to estimate the end members of a solution
}
\seealso{
Other mixing functions: \code{\link{m3_mixing_model}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
