% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiff.R
\name{stiff_transform}
\alias{stiff_transform}
\title{stiff_transform}
\usage{
stiff_transform(gdata)
}
\arguments{
\item{gdata}{A geochemical_dataset object}
}
\value{
A data.frame with the corresponding coordinates
}
\description{
Function to calculate the coordinates of a geochemical dataset in a Stiff diagram
}
\seealso{
Other stiff functions: \code{\link{ggplot_stiff}},
  \code{\link{plot_stiff}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
