% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geochemical_dataset_utilities.R
\name{[.geochemical_dataset}
\alias{[.geochemical_dataset}
\title{[.geochemical_dataset}
\usage{
\method{[}{geochemical_dataset}(x, i, ...)
}
\arguments{
\item{x}{A geochemical_dataset}

\item{i}{An index}

\item{...}{Additional parameters (for consistency purposes).}
}
\value{
This function returns a specific entry of a geochemical_dataset as a geochemical_dataset
}
\description{
Function to access an element of a geochemical_dataset.
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
