% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary.R
\name{ternary_transform}
\alias{ternary_transform}
\title{ternary_transform}
\usage{
ternary_transform(a, b, c)
}
\arguments{
\item{a, b, c}{Concentration values}
}
\value{
This function returns a data.frame with the coordinates inside a ternary diagram.
}
\description{
Function to calculate the coordinates of a point in a ternary diagram.
}
\seealso{
Other ternary functions: \code{\link{ggplot_ternary}},
  \code{\link{plot_ternary}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
