% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper.R
\name{transform_piper_data}
\alias{transform_piper_data}
\title{transform_piper_data}
\usage{
transform_piper_data(Mg, Ca, Cl, SO4, name = NULL)
}
\arguments{
\item{Mg}{A numeric vector with the concentrations of Mg}

\item{Ca}{A numeric vector with the concentrations of Ca}

\item{Cl}{A numeric vector with the concentrations of Cl}

\item{SO4}{A numeric vector with the concentrations of SO4}

\item{name}{name?}
}
\value{
A data.frame with the following variables
\itemize{
\item observation:
\item x:
\item y:
}
}
\description{
Function to transform the concentrations of Mg, Ca, Cl and SO4 to local coordinates of the Piper diagram.
}
\seealso{
Other piper functions: \code{\link{add_grid_lines_piper}},
  \code{\link{add_labels_piper}},
  \code{\link{anion_cation_transform}},
  \code{\link{ggplot_piper}},
  \code{\link{piper_classification}},
  \code{\link{piper_transform}}, \code{\link{plot_piper}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{piper functions}
