% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_utilities.R
\name{uncertainty_m3_mixing_model}
\alias{uncertainty_m3_mixing_model}
\title{uncertainty_m3_mixing_model}
\usage{
uncertainty_m3_mixing_model(gdata, end.members = NULL, lower, upper,
  nreal = 100, seed = 12345)
}
\arguments{
\item{gdata}{A geochemical_dataset object}

\item{end.members}{If provided, this is a numeric vector with the indices of the end members
to be used in the mixing calculations. If it is not specified then the end members are
defined using the convex hull of the first two principal components.}

\item{lower}{A matrix with the lower limits of the concentrations of the end.members.}

\item{upper}{A matrix with the upper limits of the concentrations of the end.members.}

\item{nreal}{The number of realizations.}

\item{seed}{Random seed.}
}
\value{
This function returns a list with the following entries:
\itemize{
\item end.members: realizations of the end.members 
\item mixing.ratios: estimated mixing ratios
}
}
\description{
Function to quantify the uncertainty in the mixing ratios due to the variability in
the concentrations of the end.members using Monte Carlo simulation.
}
\details{
This function implements a procedure designed to assess the impact of the 
compositional variability of the end members on the estimated mixing ratios. 
This procedure is based on the random generation of a predefined number of 
end member realizations where the concentration of each species is assumed 
to follow a log-normal distribution. These distributions are defined in 
terms of the upper and lower limits which corresponds to their  0.05 and 0.95 
quantiles. Then a large number of end member compostions are randomly 
generated and from each one the mixing ratios are estimated.
}
\references{
Laaksoharju, M., Skarrman, C., \& Skarrman, E. (1999). Multivariate mixing and mass
balance (M3) calculations, a new tool for decoding hydrogeochemical information.
Applied Geochemistry, 14(7), 861–871. http://doi.org/10.1016/S0883-2927(99)00024-4
}
\seealso{
Other mixing functions: \code{\link{constrained_lm}},
  \code{\link{m3_mixing_model}}, \code{\link{mix_model}},
  \code{\link{plot_m3_mixing_results}},
  \code{\link{select_end_members}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
\concept{mixing functions}
