% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary.R
\name{add_labels_ternary}
\alias{add_labels_ternary}
\title{add_labels_ternary}
\usage{
add_labels_ternary(vars, middle = TRUE, color = NULL, Size = NULL)
}
\arguments{
\item{vars}{A character vector with the names of the variables}

\item{middle}{A logical variable specifying if the labels are located in the 
central part of the ternary axis (default=TRUE)}

\item{color}{A character string specifying the color of the grid lines in a 
ternary plot}

\item{Size}{A numeric value specifying the size of the grid labels in a 
ternary plot}
}
\value{
This function returns a list with the geom_text component that specifies the 
labels of a ternary plot
}
\description{
Function to add the labels of a ternary diagram
}
\seealso{
Other ternary functions: \code{\link{add_grid_lines_ternary}},
  \code{\link{ggplot_ternary}}, \code{\link{plot_ternary}},
  \code{\link{ternary_transform}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
\concept{ternary functions}
