% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geochemical_dataset_utilities.R
\name{convert_meql}
\alias{convert_meql}
\title{convert_meql}
\usage{
convert_meql(gdata)
}
\arguments{
\item{gdata}{A data.frame with the concentrations of the major ions}
}
\value{
This function returns a list with the following entries:
\itemize{
\item anions: A matrix with the values of the normalized meqL of the anions.
\item cations: A matrix with the values of the normalized meqL of the cations.
\item meqL: A matrix with the values of the normalized meqL of the meqL.
\item mmol: A matrix with the concentration values in mmol/L.
}
}
\description{
Function to calculate the miliequivalent of a given composition.
}
\seealso{
Other base functions: \code{\link{calculate_convervative_mixing}},
  \code{\link{geochemical_dataset}},
  \code{\link{plot.geochemical_dataset}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{base functions}
