% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary.R
\name{ggplot_ternary}
\alias{ggplot_ternary}
\title{ggplot_ternary}
\usage{
ggplot_ternary()
}
\value{
This function returns the base of a ternary diagram
}
\description{
Function to create the ternary diagram
}
\seealso{
Other ternary functions: \code{\link{plot_ternary}},
  \code{\link{ternary_transform}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
