% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_utilities.R
\name{m3_mixing_model}
\alias{m3_mixing_model}
\title{m3_mixing_model}
\usage{
m3_mixing_model(gdata, end.members)
}
\arguments{
\item{gdata}{A geochemical_dataset object}

\item{end.members}{Vector with the end members of the solution}
}
\value{
This function returns a list with the following entries:
\itemize{
\item res1
\item res2
}
}
\description{
Function that implements the M# mixing model.
}
\references{
Laaksoharju, M., Skarrman, C., \& Skarrman, E. (1999). Multivariate mixing and mass
balance (M3) calculations, a new tool for decoding hydrogeochemical information.
Applied Geochemistry, 14(7), 861–871. http://doi.org/10.1016/S0883-2927(99)00024-4
}
\seealso{
Other mixing functions: \code{\link{select_end_members}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
