% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary.R
\name{plot_ternary}
\alias{plot_ternary}
\title{plot_ternary}
\usage{
plot_ternary(x, measure = c("conc", "meql"), vars = NULL,
  color = NULL, Size = NULL, labels = NULL, additional.args = NULL)
}
\arguments{
\item{x}{A geochemical_dataset object}

\item{measure}{A character variable specifying the type of measure to be used in the plot.
Currently the types supported include:
\itemize{
\item 'conc': concentrations
\item 'meql': miliequivalent
}}

\item{vars}{Character vector. Variables to use in the plot. In the case of the Durov plot, these variables
are already defined. Used only for compatibility with the plot function.}

\item{color}{Character variable that specifies the variable to color the data inside the plot.}

\item{Size}{Character variable that specifies the variable to define the size of the data inside the plot.}

\item{labels}{Character variable that specifies the labels to be used in the current plot}

\item{additional.args}{A list with additional arguments}
}
\value{
This function returns a ggplot2 object with the Durov plot.
}
\description{
Function to create a ternary plot for a geochemical_dataset
}
\seealso{
Other ternary functions: \code{\link{add_grid_lines_ternary}},
  \code{\link{add_labels_ternary}},
  \code{\link{ggplot_ternary}},
  \code{\link{ternary_transform}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
\concept{ternary functions}
