% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_utilities.R
\name{uncertainty_m3_mixing_model}
\alias{uncertainty_m3_mixing_model}
\title{uncertainty_m3_mixing_model}
\usage{
uncertainty_m3_mixing_model(gdata, end.members, lower, upper,
  nreal = 100, seed = 12345)
}
\arguments{
\item{gdata}{A geochemical_dataset object}

\item{end.members}{A geochemical_dataset object}

\item{lower}{A matrix with the lower limits of the concentrations of the end.members.}

\item{upper}{A matrix with the upper limits of the concentrations of the end.members.}

\item{nreal}{The number of realizations.}

\item{seed}{Random seed.}
}
\value{
This function returns a list with the following entries:
\itemize{
\item a
\item b
}
}
\description{
Function to quantify the uncertainty in the mixing ratios due to the variability in
the concentrations of the end.members using Monte Carlo simulation.
}
\references{
Laaksoharju, M., Skarrman, C., \& Skarrman, E. (1999). Multivariate mixing and mass
balance (M3) calculations, a new tool for decoding hydrogeochemical information.
Applied Geochemistry, 14(7), 861–871. http://doi.org/10.1016/S0883-2927(99)00024-4
}
\seealso{
Other mixing functions: \code{\link{carrera_mixing_model}},
  \code{\link{constrained_lm}},
  \code{\link{m3_mixing_model}},
  \code{\link{select_end_members}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
\concept{mixing functions}
